% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barley.yield.R
\docType{data}
\name{barley.yield}
\alias{barley.yield}
\title{Sample yield monitor data.}
\format{A SpatialPointsDataFrame  with 64698 rows and 4 variables:
\describe{
  \item{yield}{Barley yield}
  \item{speed}{Harvester speed}
  \item{distance}{Distance harvester travelled}
  \item{width}{Swath width}
}}
\usage{
data(barley.yield)
}
\description{
Sample yield monitor data.
}
\examples{
data(barley.yield)
}
\keyword{datasets}
