% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.cleaning.R
\name{clean.boundary}
\alias{clean.boundary}
\title{Remove points near paddock boundary}
\usage{
clean.boundary(yield, distance = 10, plot = FALSE,
  col = two.colors.old(start = "brown", end = "darkgreen", middle =
  "yellow"))
}
\arguments{
\item{yield}{Input yield SpatialPointsDataFrame with column 'yield'}

\item{distance}{Distance threshold from paddock boundary}

\item{plot}{TRUE/FALSE Plot yield data and detected errors}

\item{col}{colors for plotting}
}
\value{
Returns a SpatialPointsDataFrame with added column containing error flags.
}
\description{
Remove points near paddock boundary
}
\author{
Fiona Evans
}
