% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.cleaning.R
\name{clean.flow}
\alias{clean.flow}
\title{Correct flow delay errors}
\usage{
clean.flow(yield, delays = c(0:10), plot = FALSE,
  col = two.colors.old(start = "brown", end = "darkgreen", middle =
  "yellow"))
}
\arguments{
\item{yield}{Input yield SpatialPointsDataFrame with column 'yield'}

\item{delays}{Vector of delays to test}

\item{plot}{TRUE/FALSE Plot yield data and detected errors}

\item{col}{colors for plotting}
}
\value{
Returns a SpatialPointsDataFrame with added column containing error flags.
}
\description{
Correct errors due to grain flow delay. Method adapted from
Lee, D. H., et al. (2012). "Automated yield map delay identification using phase correlation methodology." 
Transactions of the ASABE 55(3): 743-752.
}
\author{
Fiona Evans
}
