% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.cleaning.R
\name{clean.geometry}
\alias{clean.geometry}
\title{Remove geometry duplicates}
\usage{
clean.geometry(yield, plot = FALSE, col = two.colors.old(start =
  "brown", end = "darkgreen", middle = "yellow"))
}
\arguments{
\item{yield}{Input yield SpatialPointsDataFrame with column 'yield'}

\item{plot}{TRUE/FALSE Plot yield data and detected errors}

\item{col}{colors for plotting}
}
\value{
Returns a SpatialPointsDataFrame with added column containing error flags.
}
\description{
Remove geometry duplicates
}
\author{
Fiona Evans
}
