% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.cleaning.R
\name{clean.lags}
\alias{clean.lags}
\title{Remove time lags}
\usage{
clean.lags(yield, quantiles = c(0.01, 0.99), plot = FALSE,
  col = two.colors.old(start = "brown", end = "darkgreen", middle =
  "yellow"))
}
\arguments{
\item{yield}{Input yield SpatialPointsDataFrame with columns 'yield' and 'distance'}

\item{quantiles}{Minimim and maximum distance quantiles}

\item{plot}{TRUE/FALSE Plot yield data and detected errors}

\item{col}{colors for plotting}
}
\value{
Returns a SpatialPointsDataFrame with added column containing error flags.
}
\description{
Remove time lags (small and large distance between points)
}
\author{
Fiona Evans
}
