% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.cleaning.R
\name{clean.outliers}
\alias{clean.outliers}
\title{Remove yield outliers}
\usage{
clean.outliers(yield, quantiles = c(0.01, 0.99), plot = FALSE,
  col = two.colors.old(start = "brown", end = "darkgreen", middle =
  "yellow"))
}
\arguments{
\item{yield}{Input yield SpatialPointsDataFrame with column 'yield'}

\item{quantiles}{Minimim and maximum yield quantiles}

\item{plot}{TRUE/FALSE Plot yield data and detected errors}

\item{col}{colors for plotting}
}
\value{
Returns a SpatialPointsDataFrame with added column containing error flags.
}
\description{
Remove yield outliers (small and large yields)
}
\author{
Fiona Evans
}
