% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.cleaning.R
\name{clean.velocity}
\alias{clean.velocity}
\title{Remove velocity outliers and rapid changes in velocity}
\usage{
clean.velocity(yield, quantiles = c(0.01, 0.99), plot = FALSE,
  col = two.colors.old(start = "brown", end = "darkgreen", middle =
  "yellow"))
}
\arguments{
\item{yield}{Input yield SpatialPointsDataFrame with columns 'yield' and 'speed'}

\item{quantiles}{Minimim and maximum velocity quantiles}

\item{plot}{TRUE/FALSE Plot yield data and detected errors}

\item{col}{colors for plotting}
}
\value{
Returns a SpatialPointsDataFrame with added column containing error flags.
}
\description{
Remove velocity outliers and rapid changes in velocity
}
\author{
Fiona Evans
}
