% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.cleaning.R
\name{clean.width}
\alias{clean.width}
\title{Remove harvest overlaps}
\usage{
clean.width(yield, plot = FALSE, col = two.colors.old(start = "brown",
  end = "darkgreen", middle = "yellow"))
}
\arguments{
\item{yield}{Input yield SpatialPointsDataFrame with columns 'yield' and 'width',}

\item{plot}{TRUE/FALSE Plot yield data and detected errors}

\item{col}{colors for plotting}
}
\value{
Returns a SpatialPointsDataFrame with added column containing error flags.
}
\description{
Remove harvest overlaps ('small' swath widths)
}
\author{
Fiona Evans
}
