% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cokrig.R
\name{cokrig.regression}
\alias{cokrig.regression}
\title{Performs local linear regressions on co-kriged OFE data.}
\usage{
cokrig.regression(data, rates)
}
\arguments{
\item{data}{SpatialPointsDataFrame output (on a grid) from global.cokrig or local.cokrig.}

\item{rates}{Vector of treatment names (first element should be the control / zero rate).}
}
\value{
Returns raster stack containing Intercept and Slope.
}
\description{
Performs local linear regressions on co-kriged OFE data. 
Fixes the Intercept to the control / zero rate.
}
\author{
Fiona Evans
}
