% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationData.R
\docType{data}
\name{data.linear.local}
\alias{data.linear.local}
\title{Simulated data with locally-varying linear response.}
\format{A SpatialPointsDataFrame  with 2400 rows and 4 variables:
\describe{
  \item{yield.min}{Minimum yield (no fertiliser applied)}
  \item{rate}{Rate of fertiliser applied}
  \item{response}{Linear rate of response to fertiliser}
  \item{yield}{Resulting yield from experiment}
}}
\usage{
data(data.linear.local)
}
\description{
Simulated on-farm experiment data with locally-varying linear response.
}
\examples{
data(data.linear.local)
head(data.linear.local)

plot.sim(data.linear.local)
}
\references{
Evans et al. 2020 MDPI Agronomy
(\href{https://address here})
}
\keyword{datasets}
