% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cokrig.R
\name{global.cokrig}
\alias{global.cokrig}
\title{Global standardised ordinary co-kriging for OFE using gstat.}
\usage{
global.cokrig(data, pred.points, yvar = "yield", xvar = "treatment",
  model = "Sph", cutoff = NULL, width = NULL)
}
\arguments{
\item{data}{SpatialPointsDataFrame containing input data.}

\item{yvar}{name of varaible to predict.}

\item{xvar}{name of treament variable.}

\item{model}{model type, e.g. "Exp", "Sph", "Gau", "Mat" (see gstat help).}

\item{cutoff}{maxium distance for variogram estimation (see variogram help).}

\item{width}{width of treatments for calculatign cutoff.}

\item{predpts}{SpatialPointsDataFrame coordinates to predict at.}
}
\value{
Returns SpatialPointsDataFrame containing predictions, variances and covariances.
}
\description{
Global standardised ordinary co-kriging for OFE using gstat.
}
\author{
Fiona Evans
}
