% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{legend.col}
\alias{legend.col}
\title{Create a color legend.}
\usage{
legend.col(lev, col = rainbow(100), ystretch = 0.5)
}
\arguments{
\item{lev}{Data used to produce colors in plot.}

\item{col}{Color map (default is rainbow).}
}
\description{
Legend for a continuous color scale in R, from 
https://aurelienmadouasse.wordpress.com/2012/01/13/legend-for-a-continuous-color-scale-in-r/
}
\keyword{color}
