% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cokrig.R
\name{local.cokrig}
\alias{local.cokrig}
\title{Local standardised ordinary co-kriging for OFE using gstat.}
\usage{
local.cokrig(data, pred.points, yvar = "yield", xvar = "treatment",
  model = "Sph", max.dist = 40)
}
\arguments{
\item{data}{SpatialPointsDataFrame containing input data.}

\item{yvar}{name of varaible to predict.}

\item{xvar}{name of treament variable}

\item{model}{model type, e.g. "Exp", "Sph", "Gau", "Mat" (see gstat help).}

\item{max.dist}{Distance threshold specifying local neighbourhood for around a prediction point,
estimating local variogram and making prediction. Must be large enough to inlcude all treatments 
at all prediction points.}

\item{predpts}{SpatialPointsDataFrame coordinates to predict at.}
}
\value{
Returns SpatialPointsDataFrame containing predictions, variances and covariances.
}
\description{
Local standardised ordinary co-kriging for OFE using gstat.
}
\author{
Fiona Evans
}
