% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cokrig.R
\name{plot.cokrig}
\alias{plot.cokrig}
\title{Plot co-kriged predictions for simulated data.}
\usage{
plot.cokrig(data, rates)
}
\arguments{
\item{data}{SpatialPointsDataFrame prediction on grid from global.cokrig / local.cokrig}

\item{rates}{Vector of treatment names / rates.}
}
\description{
Plot co-kriged predictions of simulated data.
}
\author{
Fiona Evans
}
