% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cokrig.R
\name{plot.cokrig.regression}
\alias{plot.cokrig.regression}
\title{Plot co-kriged regression results.}
\usage{
plot.cokrig.regression(data)
}
\arguments{
\item{data}{Raster stack output from cokrig.regression().}
}
\description{
Plot co-kriged regression results.
}
\author{
Fiona Evans
}
