% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationData.R
\name{plot.gwr.basic}
\alias{plot.gwr.basic}
\title{Plot results from basic GWR analysis of simulated data.}
\usage{
plot.gwr.basic(dat, model)
}
\arguments{
\item{model}{Output from GWmodel::gwr.basic}

\item{data}{SpatialPointsDataFrame (with columns 'yield.min', 'rate', 'response', 'yield') 
used as input to GWmodel::gwr.basic}
}
\description{
Plot results from basic GWR analysis of simulated data.
}
\author{
Fiona Evans
}
