% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationData.R
\name{plot.inla}
\alias{plot.inla}
\title{Plot results from INLA spatially-varying covariates analysis of simulated data.}
\usage{
plot.inla(dat, model)
}
\arguments{
\item{model}{Output from INLA::inla}

\item{data}{SpatialPointsDataFrame (with columns 'yield.min', 'rate', 'response', 'yield') 
used as input to fit INLA model}
}
\description{
Plot results from INLA spatially-varying covariates analysis of simulated data.
}
\author{
Fiona Evans
}
