% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationData.R
\name{plot.sim}
\alias{plot.sim}
\title{Plot simulated OFE SpatialPointsDataFrame (as raster).}
\usage{
plot.sim(data)
}
\arguments{
\item{data}{SpatialPointsDataFrame with columns 'yield.min', 'rate', 'response', 'yield'.}
}
\description{
Plot simulated OFE SpatialPointsDataFrame (as raster).
}
\examples{
data(data.linear.local)

plot.sim(data.linear.local)

}
\author{
Fiona Evans
}
