% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{qfilter}
\alias{qfilter}
\title{Filter data to remove low and high outliers}
\usage{
qfilter(x, low = 0.01, high = 0.99)
}
\arguments{
\item{x}{Input data.}

\item{low}{Value of low quantile (default = 0.01).}

\item{high}{Value of high quantile (default = 0.99).}
}
\value{
Returns a vector of station ids, in order of increasing distance.
}
\description{
Filter data to removel ow and high outliers
}
\author{
Fiona Evans
}
