% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationData.R
\name{sim.to.raster}
\alias{sim.to.raster}
\title{Convert simulated OFE SpatialPointsDataFrame to raster stack.}
\usage{
sim.to.raster(data)
}
\arguments{
\item{data}{SpatialPointsDataFrame with columns 'yield.min', 'rate', 'response', 'yield'.}
}
\value{
Returns a raster stack
}
\description{
Convert simulated OFE SpatialPointsDataFrame to raster stack.
}
\author{
Fiona Evans
}
