% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{spatial.interp.inla}
\alias{spatial.interp.inla}
\title{Interpolate to a grid using R-INLA}
\usage{
spatial.interp.inla(dat, varname, width, crop = F, plot = F,
  col = two.colors.old(start = "brown", end = "darkgreen", middle =
  "yellow"))
}
\arguments{
\item{dat}{Input SpatialPointsDataFrame with Easting-Northing projection CRS("+proj=utm")}

\item{varname}{Name of col in dat to interpolate}

\item{width}{Width of harvester / spreader}

\item{crop}{TRUE/FALSE Crop concave hull}

\item{plot}{TRUE/FALSE Plot interpolated data}

\item{col}{colors for plotting}
}
\value{
Returns a raster.
}
\description{
Interpolate to a grid using R-INLA
Expects columns in input SpatialPointsDataFrame: yield, speed, width, distance
}
\author{
Fiona Evans
}
