% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cokrig.R
\name{treatment.distance}
\alias{treatment.distance}
\title{Minimum distance covering all treatments.}
\usage{
treatment.distance(data, yvar = "yield", xvar = "treatment")
}
\arguments{
\item{data}{SpatialPointsDataFrame containing input data.}

\item{yvar}{name of varaible to predict.}

\item{xvar}{name of treament variable.}
}
\value{
Returns numeric distance.
}
\description{
Determines minimum distance that includes all treatments from all data points.
}
\author{
Fiona Evans
}
